renamed vd helper utils module
