+++ mmif/utils/__init__.py
@@ -1,1 +1,1 @@
-from mmif.utils import video_document
+from mmif.utils import video_document_helper
